requirements.txt
pycryptodome==3.18.0
pywin32==306